Cant figure out the "Add" visitor onpress function
I've set up the if statements for the card states. Default is false (which shows the Add Visitor card ui). When switched to true it switches the ui to the active visitor card. Im not sure how to go about adding the onpress function when a user clicks "Add" on the popup form.