Visitor Functions
Functions work for the most part but has bugs.

Bug 1 : Parking duration is working and sets the selected duration but the picker ui slides back to 2hrs everytime (like before)
Bug 2: Extend does not add to the current duration, but sets it to the selected time. (Instead of adding +2 hours to 4 hours, it sets the time to 2hours)
Bug 3: Remove button works but does not remove the correct card. Always removes the last card first. Not sure what kind of if statement to put for that.

Any tips would be helpful!