Database Setup Done
Added functions in <Main /> for communicating with DB, including ADDING, REMOVING, EXTENDING new visitors and VIEWING history.
Pinning a past visitor in progress.
Added folder "server" that includes PHP files that need to be put on the server and a file "mysql" --- commands for building tables in mysql in terminal.